Add Nav Button component
Changed a couple of classes to function components when state was not necessary.
Nav Button forced rewrite of the TextInputField component because state was not being passed around, but instead using the prop passed down from the parent which affected state.

Timer Screen has two callbacks which are separated out for clearer functionality and allowing state change to not affect both state properties.