Timer Display state transition added
Timer Display starts at a smaller dimension. When state changes to the timer having started, it does a ease transition into a larger state.

This is done using ems and font size. Since these digits aren't actual fonts, the best way to transition everything is to use the font size value to affect how a em dimension is displayed.

1em of 16px is larger than 1em of 5px. The a CSS transition smoothly eases the size differences.